requirements.txt
openai==1.54.3
python-dotenv==1.0.1
